+++ src/Main.java
@@ -1,23 +1,23 @@
-import service.AttendanceService;
-import service.StudentService;
+import service.AttendanceServiceOld;
+import service.StudentServiceOld;
 import ui.MainMenu;
 
 public class Main {
     public static void main(String[] args) {
 
         //Services
-        StudentService studentService = new StudentService();
-        AttendanceService attendanceService = new AttendanceService(studentService.students);
-        MainMenu mainMenu = new MainMenu(studentService, attendanceService);
+        StudentServiceOld studentServiceOld = new StudentServiceOld();
+        AttendanceServiceOld attendanceServiceOld = new AttendanceServiceOld(studentServiceOld.students);
+        MainMenu mainMenu = new MainMenu(studentServiceOld, attendanceServiceOld);
         mainMenu.printMainMenu();
 
         // Listeye proje başlatıldığı zaman default 10 adet talebe ekler.
         // Bir sonraki talebe 11. sıradan başlar.
-        studentService.addStudent("Veli", "Çam");
-        studentService.addStudent("Abdurrahman", "Kutlu");
-        studentService.addStudent("Emre", "Yavuz");
-        studentService.addStudent("Kaan", "Koca");
-        studentService.addStudent("Enes Bahadır", "Yıldırım");
+        studentServiceOld.addStudent("Veli", "Çam");
+        studentServiceOld.addStudent("Abdurrahman", "Kutlu");
+        studentServiceOld.addStudent("Emre", "Yavuz");
+        studentServiceOld.addStudent("Kaan", "Koca");
+        studentServiceOld.addStudent("Enes Bahadır", "Yıldırım");
 //        studentService.addStudent("Enver", "Yıldırım");
 //        studentService.addStudent("Yasin", "Büzgülü");
 //        studentService.addStudent("Bektaş", "Işık");
