+++ src/ui/StudentMenu.java
@@ -2,19 +2,19 @@ package ui;
 
 import domain.Student;
 import org.jetbrains.annotations.NotNull;
-import service.StudentService;
+import service.StudentServiceOld;
 
 import java.util.List;
 import java.util.Scanner;
 
 public class StudentMenu {
-    private final StudentService studentService;
+    private final StudentServiceOld studentServiceOld;
     private final Scanner input = new Scanner(System.in);
     public List<Student> students;
 
-    public StudentMenu(@NotNull StudentService studentService) {
-        this.students = studentService.students;
-        this.studentService = studentService;
+    public StudentMenu(@NotNull StudentServiceOld studentServiceOld) {
+        this.students = studentServiceOld.students;
+        this.studentServiceOld = studentServiceOld;
     }
 
     public void printStudentMenu() {
@@ -35,16 +35,16 @@ public class StudentMenu {
             String choice = input.nextLine();
             switch (choice) {
                 case "1":
-                    studentService.addStudent();
+                    studentServiceOld.addStudent();
                     break label;
                 case "2":
-                    studentService.printStudents();
+                    studentServiceOld.printStudents();
                     break label;
                 case "3":
-                    studentService.updateStudent();
+                    studentServiceOld.updateStudent();
                     break label;
                 case "4":
-                    studentService.deleteStudent();
+                    studentServiceOld.deleteStudent();
                     break label;
                 case "5":
                     break label;
